defrag: gate serde as native-only and extract shared diagnostic tree walk
